refactor[#15] : features/post 레이어 article 로 슬라이스명 변경
